Rename LexemeBuilder to LexemeAnalyzer
