+++ packages/web/src/lexemeAnalyzer/types.ts
