--- packages/web/src/lexemeBuilder/index.ts
@@ -1,2 +1,0 @@
-export * from './LexemeBuilder';
-export * from './types';
